% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{coocurrenceDCHFO}
\alias{coocurrenceDCHFO}
\title{Title  compute cooccurrence index of DC Shift and HFO}
\usage{
coocurrenceDCHFO(testDce, testHfo)
}
\arguments{
\item{testDce}{Numeric. DC shift is detected on electrodes with binary value 1}

\item{testHfo}{Numeric. Hfo shift is detected on electrodes with binary value 1}
}
\value{
electrode indices of coocurrence
}
\description{
Title  compute cooccurrence index of DC Shift and HFO
}
\examples{
indexCooc<-cooccurrenceDCHfo(testDce,testHfo)
}
