% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_DCShift.r
\name{nrow,DCShift-method}
\alias{nrow,DCShift-method}
\alias{ncol,DCShift-method}
\title{Get the number of rows or columns of a DCShift object}
\usage{
\S4method{nrow}{DCShift}(x)

\S4method{ncol}{DCShift}(x)
}
\arguments{
\item{x}{A DCShift object}
}
\description{
Get the number of rows or columns of a DCShift object
}
