% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_HFOMetrics.r
\name{nrow,HFOMetrics-method}
\alias{nrow,HFOMetrics-method}
\alias{ncol,HFOMetrics-method}
\title{Get the number of rows or columns of a HFOMetrics object}
\usage{
\S4method{nrow}{HFOMetrics}(x)

\S4method{ncol}{HFOMetrics}(x)
}
\arguments{
\item{x}{A HFOMetrics object}
}
\description{
Get the number of rows or columns of a HFOMetrics object
}
