% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_DCShift.r
\name{show,DCShift-method}
\alias{show,DCShift-method}
\title{Print the DCShift object}
\usage{
\S4method{show}{DCShift}(object)
}
\arguments{
\item{object}{A DCShift object}
}
\description{
Print the DCShift object
}
