% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_HFOMetrics.r
\name{show,HFOMetrics-method}
\alias{show,HFOMetrics-method}
\title{Print the HFOMetrics object}
\usage{
\S4method{show}{HFOMetrics}(object)
}
\arguments{
\item{object}{A HFOMetrics object}
}
\description{
Print the HFOMetrics object
}
