% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_DCShift.r
\name{[,DCShift,ANY,ANY,ANY-method}
\alias{[,DCShift,ANY,ANY,ANY-method}
\title{Subset a DCShift object}
\usage{
\S4method{[}{DCShift,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{A DCShift object}

\item{i}{A logical vector or a numeric vector of indices to subset the electrodes}

\item{j}{A logical vector or a numeric vector of indices to subset the time windows}

\item{...}{Additional arguments (not used)}

\item{drop}{Additional arguments (not used)}
}
\description{
Subset a DCShift object
}
