% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_HFOMetrics.r
\name{[,HFOMetrics,ANY,ANY,ANY-method}
\alias{[,HFOMetrics,ANY,ANY,ANY-method}
\title{Subset a HFOMetrics object}
\usage{
\S4method{[}{HFOMetrics,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{A HFOMetrics object}

\item{i}{A logical vector or a numeric vector of indices to subset the electrodes}

\item{j}{A logical vector or a numeric vector of indices to subset the time windows}

\item{...}{Additional arguments (not used)}

\item{drop}{Additional arguments (not used)}
}
\description{
Subset a HFOMetrics object
}
