% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{analyze_DCShift}
\alias{analyze_DCShift}
\title{compute DC shift analysis}
\usage{
analyze_DCShift(
  epoch,
  fs = 1000,
  thresholdDc = 0.5,
  lengthDc = 3,
  thresholdEndStartDc = 0.001
)
}
\arguments{
\item{epoch}{Matrix or Epoch object. iEEG data matrix or Epoch object. If matrix, the row names are the electrode names and the column names are the time points}

\item{fs}{Numeric. frequency of signal iEEG acquisition}

\item{thresholdDc.}{Numeric. Threshold amplitude for DC shift}

\item{lengthDC.}{Numeric. Threshold in length of DC shift in seconds}

\item{thresholdEndStartDc.}{Numeric. Threshold to detect the start of the DC shift}
}
\value{
A DC shift analysis object
}
\description{
compute DC shift analysis
}
\examples{
data("pt01EcoG")
ts<-pt01EcoG[,20001:50001]
fs=1000
sozIndex <- attr(pt01EcoG, "sozIndex")
windowParams<-c(0.25,0.1)
epoch <- Epoch(ts)
visuIEEGData(epoch)
}
